Added timed auctions and fees for auctions and storage

+++ WebInterface/scripts/newAuction.php
@@ -7,14 +7,18 @@
 	$user = $_SESSION['User'];
 	$canSell = $_SESSION['canSell'];
 	if ($canSell == false){
-		header("Location: ../myauctions.php?error=7");
+		$_SESSION['error'] = 'You do not have permission to sell items.';
+		header("Location: ../myauctions.php");
 	}
 	require 'config.php';
 	require 'itemInfo.php';
+	require_once '../classes/EconAccount.php';
 	if ($useTwitter == true){require_once 'twitter.class.php';}
 	$itemNameFull = explode(":", $_POST['Item']);
 	$sellName = $itemNameFull[0];
 	$sellDamage = $itemNameFull[1];
+	$player = new EconAccount($user, $useMySQLiConomy, $iConTableName);
+	$marketPrice = getMarketPrice($sellName, $sellDamage);
 	$sellPrice = round($_POST['Price'], 2);
 	$itemFullName = getItemName($sellName, $sellDamage);
 	if ($sellPrice > $maxSellPrice){ $sellPrice == $maxSellPrice; }
@@ -24,11 +28,13 @@
 	$queryItems=mysql_query("SELECT * FROM WA_Items WHERE player='$user' AND name='$sellName' AND damage='$sellDamage'");
 	list($id, $itemName, $itemDamage, $itemOwner, $itemQuantity)= mysql_fetch_row($queryItems);
     if ($sellQuantity <= 0){
-		header("Location: ../myauctions.php?error=5");
+		$_SESSION['error'] = 'Quantity was not a valid number.';
+		header("Location: ../myauctions.php");
 	}
 	if ($sellPrice <= 0)
 	{
-		header("Location: ../myauctions.php?error=6");
+		$_SESSION['error'] = 'Price was not a valid number.';
+		header("Location: ../myauctions.php");
 	}
 	else{
 		if (is_numeric($sellPrice)){	
@@ -37,35 +43,48 @@
 				if ($itemQuantity >= $sellQuantity)
 				{
 					$itemsLeft = $itemQuantity - $sellQuantity;
-					if ($sellQuantity > 0)
+					$itemFee = (($marketPrice/100)*$auctionFee);
+					if ($player->money >= $itemFee){
+						if ($sellQuantity > 0)
+						{
+							$timeNow = time();
+							$player->money = $player->money - $itemFee;
+							$player->saveMoney($useMySQLiConomy, $iConTableName);
+							$itemQuery = mysql_query("INSERT INTO WA_Auctions (name, damage, player, quantity, price, created) VALUES ('$sellName', '$sellDamage', '$user', '$sellQuantity', '$sellPrice', '$timeNow')");
+						}
+						if ($itemsLeft == 0)
+						{
+							$itemDelete = mysql_query("DELETE FROM WA_Items WHERE id='$id'");
+						}
+						else
+						{
+							$itemUpdate = mysql_query("UPDATE WA_Items SET quantity='$itemsLeft' WHERE id='$id'");
+						}
+						if ($useTwitter == true){
+							$twitter = new Twitter($consumerKey, $consumerSecret, $accessToken, $accessTokenSecret);
+							$twitter->send('[WA] Auction Created: '.$user.' is selling '.$sellQuantity.' x '.$itemFullName.' for '.$currencyPrefix.$sellPrice.$currencyPostfix.' each. At '.date("H:i:s").'. '.$shortLinkToAuction.' #webauction');
+						}
+						$_SESSION['success'] = "You auctioned $sellQuantity $itemFullName for ".$currencyPrefix.$totalPrice.$currencyPostfix.", the fee was ".$currencyPrefix.$itemFee.$currencyPostfix;
+						header("Location: ../myauctions.php");
+					}else
 					{
-						$itemQuery = mysql_query("INSERT INTO WA_Auctions (name, damage, player, quantity, price) VALUES ('$sellName', '$sellDamage', '$user', '$sellQuantity', '$sellPrice')");
+					  $_SESSION['error'] = 'Fee cost '.$currencyPrefix.$itemFee.$currencyPostfix.', you did not have enough money.';
+					  header("Location: ../myauctions.php");
 					}
-					if ($itemsLeft == 0)
-					{
-						$itemDelete = mysql_query("DELETE FROM WA_Items WHERE id='$id'");
-					}
-					else
-					{
-						$itemUpdate = mysql_query("UPDATE WA_Items SET quantity='$itemsLeft' WHERE id='$id'");
-					}
-					if ($useTwitter == true){
-						$twitter = new Twitter($consumerKey, $consumerSecret, $accessToken, $accessTokenSecret);
-						$twitter->send('[WA] Auction Created: '.$user.' is selling '.$sellQuantity.' x '.$itemFullName.' for '.$currencyPrefix.$sellPrice.$currencyPostfix.' each. At '.date("H:i:s").'. '.$shortLinkToAuction.' #webauction');
-					}
+				}else
+				{
+				    $_SESSION['error'] = 'You do not have enough of that item.';
 					header("Location: ../myauctions.php");
-				}else{
-					header("Location: ../myauctions.php?error=2");
 				}
-			}
-			else
+			}else
 			{
-				header("Location: ../myauctions.php?error=4");
+				$_SESSION['error'] = 'Quantity was not an integer.';
+				header("Location: ../myauctions.php");
 			}
-		}
-		else
+		}else
 		{
-			header("Location: ../myauctions.php?error=3");
+			$_SESSION['error'] = 'Price was not an integer.';
+			header("Location: ../myauctions.php");
 		}
 	}
